Move summary related code into it's own package
